+++ packages/hub/schema.graphql
@@ -211,6 +211,7 @@ type ModelExportRevisionsConnectionEdge {
 
 type ModelRevision implements Node {
   author: User
+  builds: [ModelRevisionBuild!]!
   comment: String!
   content: ModelContent!
   createdAtTimestamp: Float!
@@ -219,7 +220,22 @@ type ModelRevision implements Node {
   id: ID!
   model: Model!
   relativeValuesExports: [RelativeValuesExport!]!
-  runs: [ModelRevisionRun!]!
+}
+
+type ModelRevisionBuild implements Node {
+  errors: [String!]!
+  id: ID!
+  modelRevision: ModelRevision!
+}
+
+type ModelRevisionBuildConnection {
+  edges: [ModelRevisionBuildEdge!]!
+  pageInfo: PageInfo!
+}
+
+type ModelRevisionBuildEdge {
+  cursor: String!
+  node: ModelRevisionBuild!
 }
 
 type ModelRevisionConnection {
@@ -242,24 +258,6 @@ union ModelRevisionForRelativeValuesResult = BaseError | NotFoundError | Relativ
 input ModelRevisionForRelativeValuesSlugOwnerInput {
   owner: String!
   slug: String!
-}
-
-type ModelRevisionRun implements Node {
-  endedAtTimestamp: Float!
-  errors: [String!]!
-  id: ID!
-  modelRevision: ModelRevision!
-  startedAtTimestamp: Float!
-}
-
-type ModelRevisionRunConnection {
-  edges: [ModelRevisionRunEdge!]!
-  pageInfo: PageInfo!
-}
-
-type ModelRevisionRunEdge {
-  cursor: String!
-  node: ModelRevisionRun!
 }
 
 type ModelsByVersion {
