+++ packages/hub/src/graphql/schema.ts
@@ -1,7 +1,7 @@
 import "./errors/BaseError";
 import "./errors/NotFoundError";
 import "./errors/ValidationError";
-import "./types/ModelRevisionRun";
+import "./types/ModelRevisionBuild";
 import "./queries/globalStatistics";
 import "./queries/group";
 import "./queries/groups";
