+++ packages/hub/src/graphql/types/ModelRevision.ts
@@ -35,7 +35,7 @@ export const ModelRevision = builder.prismaNode("ModelRevision", {
     relativeValuesExports: t.relation("relativeValuesExports"),
     exports: t.relation("exports"),
     model: t.relation("model"),
-    runs: t.relation("runs"),
+    builds: t.relation("builds"),
     author: t.relation("author", { nullable: true }),
     comment: t.exposeString("comment"),
     content: t.field({
